fix: move all constants to root path
