--- src/components/WordBook/types.ts
@@ -1,21 +1,0 @@
-export interface IGroup {
-  label: string;
-  linkAddress: string;
-  color: COLORS;
-  background: COLORS;
-}
-
-export enum COLORS {
-  RED = '#DD4132',
-  YELLOW = '#EFC050',
-  ORANGE = '#FE840E',
-  GREEN = '#00A170',
-  BLUE = '#3F69AA',
-  PURPLE = '#6C244C',
-  BG_RED = '#fbebe9',
-  BG_YELLOW = '#fdf7e8',
-  BG_ORANGE = '#fff2e6',
-  BG_GREEN = '#e6fff7',
-  BG_BLUE = '#ecf1f8',
-  BG_PURPLE = '#f9ecf3',
-}
